created utils and app
By creating app.js, I separated the transactions and made them more readable. I prepared the basis for creating middleware with the utils folder.
