+++ server/app.js
@@ -1,0 +1,34 @@
+const express = require('express')
+const app = express()
+const bodyParser = require('body-parser')
+const mongoose = require('mongoose')
+const cors = require('cors')
+
+const config = require('./utils/config')
+const logger = require('./utils/logger')
+
+const indexRouter = require('./routes/index')
+const postRouter = require('./routes/posts')
+
+app.use(bodyParser.json({ limit: '30mb', extended: true }))
+app.use(bodyParser.urlencoded({ limit: '30mb', extended: true }))
+app.use(cors())
+
+mongoose
+  .connect(config.MONGODB_URI, {
+    useNewUrlParser: true,
+    useUnifiedTopology: true,
+    useFindAndModify: false,
+    useCreateIndex: true,
+  })
+  .then(() => {
+    logger.info('connected to MongoDB')
+  })
+  .catch((error) => {
+    logger.error('error connecting to MongoDB', error.message)
+  })
+
+app.use('/', indexRouter)
+app.use('/posts', postRouter)
+
+module.exports = app
