+++ server/index.js
@@ -1,30 +1,7 @@
-if (process.env.NODE_ENV !== 'production') {
-  require('dotenv').config()
-}
+const app = require('./app')
+const config = require('./utils/config')
+const logger = require('./utils/logger')
 
-const express = require('express')
-const app = express()
-const bodyParser = require('body-parser')
-const mongoose = require('mongoose')
-const cors = require('cors')
-
-const indexRouter = require('./routes/index')
-const postRouter = require('./routes/posts')
-
-app.use(bodyParser.json({ limit: '30mb', extended: true }))
-app.use(bodyParser.urlencoded({ limit: '30mb', extended: true }))
-app.use(cors())
-
-mongoose.connect(process.env.MONGODB_URI, {
-  useNewUrlParser: true,
-  useUnifiedTopology: true,
+app.listen(config.PORT, () => {
+  logger.info(`Server running on port: ${config.PORT}`)
 })
-
-const db = mongoose.connection
-db.on('error', (error) => console.error(error))
-db.once('open', () => console.log('Connected to Mongoose'))
-
-app.use('/', indexRouter)
-app.use('/posts', postRouter)
-
-app.listen(process.env.PORT || 5000)
