+++ server/utils/logger.js
@@ -1,0 +1,12 @@
+const info = (...params) => {
+  console.log(...params)
+}
+
+const error = (...params) => {
+  console.log(...params)
+}
+
+module.exports = {
+  info,
+  error,
+}
